+++ client/src/components/IconButton.svelte
