--- client/src/components/icon-button/index.ts
@@ -1,3 +1,0 @@
-import IconButtonSvelte from "./IconButton.svelte";
-
-export default IconButtonSvelte;
